Zuladung zu float geändert

+++ Fahrzeugverwaltung/Klassen/LKW.cs
@@ -7,11 +7,11 @@ namespace Fahrzeugverwaltung
     {
         //festlegen der Variablen der Klasse LKW
         private int achsenAnzahl;
-        private int zuladung;
+        private float zuladung;
 
         //Konstruktor anlegen
         //erbt Konstruktor von der Basisklasse Fahrzeug und erweitert diesen um die Achsenanzahl und die Zuladung
-        public LKW(String aHersteller, String aModell, String aKennzeichen, int aErstzulassung, float aAnschaffungspreis, int aAchsenanzahl, int aZuladung)
+        public LKW(String aHersteller, String aModell, String aKennzeichen, int aErstzulassung, float aAnschaffungspreis, int aAchsenanzahl, float aZuladung)
         : base(aHersteller, aModell, aKennzeichen, aErstzulassung, aAnschaffungspreis)
         {
             achsenAnzahl = aAchsenanzahl;
@@ -20,6 +20,6 @@ namespace Fahrzeugverwaltung
 
         //Getter und Setter der Klasse LKW
         public int Achsenanzahl { get { return achsenAnzahl; } set { achsenAnzahl = value; } }
-        public int Zuladung { get { return zuladung; } set { zuladung = value; } }
+        public float Zuladung { get { return zuladung; } set { zuladung = value; } }
     }
 }
